Split out form validation to form-access and bg-form-submit
