+++ static/js/src/core.js
@@ -1,5 +1,6 @@
 import { createNav } from "@canonical/global-nav";
 import { cookiePolicy } from "@canonical/cookie-policy";
+import { assignMarketoBackgroundSubmit } from "./bg-form-submit";
 
 // Initalise the global navigation.
 createNav({
@@ -8,5 +9,7 @@ createNav({
 });
 
 // Initalise the cookie policy notification.
-
 cookiePolicy();
+
+// Add background submit functionality to Marketo forms
+assignMarketoBackgroundSubmit();
