+++ webpack.config.entry.js
@@ -16,7 +16,8 @@ module.exports = {
     "./static/js/src/dynamic-forms.js",
     "./static/js/src/core.js",
     "./static/js/src/navigation.js",
-    "./static/js/src/form-validation.js",
+    "./static/js/src/form-access.js",
+    "./static/js/src/utils/bg-form-submit.js",
     "./static/js/src/scratch.js",
     "./static/js/src/smartquotes.js",
   ],
